+++ conan/tools/gnu/autotoolstoolchain.py
@@ -1,13 +1,12 @@
 import json
 
 from conan.tools import CONAN_TOOLCHAIN_ARGS_FILE
-from conan.tools._compilers import architecture_flag, build_type_flags
+from conan.tools._compilers import architecture_flag, build_type_flags, cppstd_flag
 from conan.tools.apple.apple import apple_min_version_flag, to_apple_arch, \
     apple_sdk_path
 from conan.tools.cross_building import cross_building, get_cross_building_settings
 from conan.tools.env import Environment
 from conan.tools.files import save
-from conan.tools.gnu.get_cppstd import cppstd_flag
 from conan.tools.gnu.get_gnu_triplet import _get_gnu_triplet
 from conans.tools import args_to_string
